# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@supchat.io
    name: Support API SupChat
  description: API pour le service de messagerie SupChat
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SupChat API
  version: "1.0"
paths:
...
  /api/workspaces:
    get:
      consumes:
      - application/json
      description: Récupère tous les espaces de travail auxquels l'utilisateur appartient
      produces:
      - application/json
      responses:
        "200":
          description: Liste des espaces de travail de l'utilisateur
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "401":
          description: Utilisateur non authentifié
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la récupération des espaces de travail
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Espaces de travail de l'utilisateur
      tags:
      - workspace
    post:
      consumes:
      - application/json
      description: Crée un nouvel espace de travail avec l'utilisateur connecté comme
        propriétaire
      parameters:
      - description: Informations de l'espace de travail
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Espace de travail créé avec succès
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Requête invalide ou données incomplètes
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Utilisateur non authentifié
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la création de l'espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Création d'un espace de travail
      tags:
      - workspace
  /api/workspaces/{workspace_id}:
    get:
      consumes:
      - application/json
      description: Récupère les informations de base d'un espace de travail spécifique
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Informations de l'espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID de l'espace de travail manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Utilisateur non autorisé dans cet espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la récupération de l'espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Détails d'un espace de travail
      tags:
      - workspace
    put:
      consumes:
      - application/json
      description: Modifie le nom et le sujet d'un espace de travail
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Informations à mettre à jour
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Informations mises à jour avec succès
        "400":
          description: ID de l'espace de travail manquant ou requête invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Utilisateur non autorisé dans cet espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la mise à jour des informations
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mise à jour des informations de l'espace de travail
      tags:
      - workspace
...
  /api/workspaces/{workspace_id}/type:
    put:
      consumes:
      - application/json
      description: Modifie le type d'un espace de travail (public/privé)
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Nouveau type d'espace de travail
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Type mis à jour avec succès
        "400":
          description: ID de l'espace de travail manquant ou requête invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Utilisateur non autorisé dans cet espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la mise à jour du type
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mise à jour du type d'espace de travail
      tags:
      - workspace
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"